invoice pdf refactor

+++ api/merchant/invoice/invoice.go
@@ -44,7 +44,7 @@ type ListReq struct {
 	Status        []int  `json:"status" dc:"The status of invoice, 1-pending｜2-processing｜3-paid | 4-failed | 5-cancelled" `
 	AmountStart   int64  `json:"amountStart" dc:"The filter start amount of invoice" `
 	AmountEnd     int64  `json:"amountEnd" dc:"The filter end amount of invoice" `
-	UserId        int    `json:"userId" dc:"The filter userid of invoice" `
+	UserId        uint64 `json:"userId" dc:"The filter userid of invoice" `
 	SendEmail     string `json:"sendEmail" dc:"The filter email of invoice" `
 	SortField     string `json:"sortField" dc:"Filter，em. invoice_id|gmt_create|gmt_modify|period_end|total_amount，Default gmt_modify" `
 	SortType      string `json:"sortType" dc:"Sort，asc|desc，Default desc" `
